+++ packages/gestalt-codemods/generic-codemods/__testfixtures__/modifyPropValue/errorSpreadProp.input.js
@@ -1,0 +1,12 @@
+import { Box } from 'gestalt';
+
+export default function TestComp() {
+  const props = {}
+  return (
+    <Box {...props} width={400}>
+      <Box {...props} >
+        <Box/>
+      </Box>
+    </Box>
+  );
+}
