+++ packages/gestalt-codemods/generic-codemods/__testfixtures__/modifyPropValue/errorSpreadProp.output.js
@@ -1,0 +1,11 @@
+import { Box } from 'gestalt';
+
+export default function TestComp() {
+  return (
+    <Box width={400} variant="error">
+      <Box variant="error">
+        <Box variant="error" />
+      </Box>
+    </Box>
+  );
+}
