+++ packages/gestalt-codemods/generic-codemods/__testfixtures__/modifyPropValue/renamePropValueTrueBoolean.input.js
@@ -1,0 +1,11 @@
+import { Box, Button } from 'gestalt';
+
+export default function TestComp() {
+  return (
+    <Box width={400} fit>
+      <Box fit/>
+      <Box fit/>
+      <Button/>
+    </Box>
+  );
+}
