+++ packages/gestalt-codemods/generic-codemods/__testfixtures__/modifyPropValue/renamePropValueTrueBoolean.output.js
@@ -1,0 +1,11 @@
+import { Box, Button } from 'gestalt';
+
+export default function TestComp() {
+  return (
+    <Box width={400} left={false}>
+      <Box left={false}/>
+      <Box left={false}/>
+      <Button/>
+    </Box>
+  );
+}
